Rao1999 hpc (#135)
* Enhance documentation for predictive coding model

Expanded the documentation for the predictive coding model, detailing the construction of neural and synaptic components, process dynamics, and training procedures.

* Add files via upload

* Update image path for GEC in documentation

* Revise PC model documentation formatting

Updated headings and formatting for clarity in PC model documentation.

* Change header levels for PC model training sections

* Add files via upload

* Update PC model training section with input image details

Added explanation about the input image for the PC model training.

* Update pc_rao_ballard1999.md

* Delete docs/images/museum/hgpc/Patch_input.png

* Add files via upload

* Fix image source and enhance PC model description

Updated image source and adjusted description for clarity.

* Refactor PC model training sections in documentation

Removed the section for training the PC model on the full image and added a reference to it in the patched image section.